Create taxalgorithm
Parameters of the problem:
can add or subtract up to $0.01 to the price of a product.
This is a bruteforce way of calculating all the possible prices of the 12 products on the menu. 
Caveat: it takes about an hr to run all the calculations, admittedly not the most efficient way, but guarantees reaching the global optima. 
Written in Matlab.